+++ nautilus_trader/model/c_enums/order_type.pyx
@@ -27,12 +27,14 @@ cdef class OrderTypeParser:
         elif value == 4:
             return "STOP_LIMIT"
         elif value == 5:
-            return "MARKET_IF_TOUCHED"
+            return "MARKET_TO_LIMIT"
         elif value == 6:
-            return "LIMIT_IF_TOUCHED"
+            return "MARKET_IF_TOUCHED"
         elif value == 7:
-            return "TRAILING_STOP_MARKET"
+            return "LIMIT_IF_TOUCHED"
         elif value == 8:
+            return "TRAILING_STOP_MARKET"
+        elif value == 9:
             return "TRAILING_STOP_LIMIT"
         else:
             raise ValueError(f"value was invalid, was {value}")
@@ -47,6 +49,8 @@ cdef class OrderTypeParser:
             return OrderType.STOP_MARKET
         elif value == "STOP_LIMIT":
             return OrderType.STOP_LIMIT
+        elif value == "MARKET_TO_LIMIT":
+            return OrderType.MARKET_TO_LIMIT
         elif value == "MARKET_IF_TOUCHED":
             return OrderType.MARKET_IF_TOUCHED
         elif value == "LIMIT_IF_TOUCHED":